obscure owners api key for transitfeeds.com
The security of releasing an API key for a gtfs hosting website might be
pretty low in comparison to other api keys but I thought it would be an
improvement to remove it from the source so others couldn't use it or
draw the wrong attention to the project.  I also added a file to the
gitignore where each developer can put their own key and read it in
where it was called before.  Also sneaking in the installation change
into the Rmd file to match the change in the md file.