break up get_sldf; try sf and sp df's
there's a lot of great error checking and work in the get_sldf function. however, most of it is hard to get at unless you're trying to put things directly onto a map. i've found this package useful for a number of reasons, mainly in assisting in some work focused on geospatial questions. as a means of learning more about how the get_sldf function was building the relationship between routes, trips, and shapes, i broke it up into a number of functions. it still seems to work as intended this way. hopefully these components will allow me to swap out, for example, the sp classes for the sf classes, as suggested in this issue https://github.com/ropensci/gtfsr/issues/24