clarify and document date_service
this is an attempt to resolve https://github.com/r-transit/tidytransit/issues/27. ran devtools::check() and filled in global variables for rlang, added param definition. renamed a function to make it clearer, and made it return a df rather than the whole gtfs obj. that way we can use it elsewhere later (for example automatically in read_gtfs). made set_hms_time internal until we have a solid example on how to use it.